move commitment proof to types
