+++ store/types/proof.go
@@ -1,4 +1,4 @@
-package iavl
+package types
 
 import (
 	"errors"
